generic test suite for Emojione +JS implementation
The file 'lib/validate.yml' contains a cross-platform definition
of the test cases for the Emojione library, in a similar fashion
to Twitter's (https://github.com/twitter/twitter-text-conformance).

This allows different implementations of Emojione, such as the php
and js one, to stay in a consistent, stable and interoperable state.

The file 'lib/js/js-test' implements the JavaScript tests for Emojione.
It is a small Perl program that parses 'lib/validate.yml' and
auto-generates a file called 'lib/js/test.html', containing QUnit
JavaScript tests for Emojione, then fires up the tests in the web browser.